dev-knowledge-base-requirements.txt
# Knowledge Base Requirements
# For Qdrant-based RAG system

# Core dependencies
qdrant-client==1.12.1
langchain==0.3.26
langchain-qdrant==0.2.4
langchain-openai==0.3.0
langchain-community==0.3.21

# OpenAI
openai==1.59.6

# Document processing
pypdf2==3.0.1
GitPython==3.1.43
beautifulsoup4==4.12.3

# Database
pymongo==4.6.1
beanie==1.23.6
motor==3.3.2

# Utilities
python-dotenv==1.0.1
click==8.1.7
rich==13.9.4
structlog==24.1.0

# Text processing
tiktoken==0.8.0